+++ HealthCareSystem/src/main/java/com/cg/healthcare/controller/TestController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.cg.healthcare.entities.DiagnosticCenter;
 import com.cg.healthcare.entities.DiagnosticTest;
 import com.cg.healthcare.exception.DataNotFoundInDataBase;
+import com.cg.healthcare.exception.ForBiddenException;
 import com.cg.healthcare.service.ITestService;
 
 @RestController
@@ -21,24 +22,33 @@ import com.cg.healthcare.service.ITestService;
 public class TestController {
 	@Autowired
 	ITestService testService;
+	
+	@Autowired
+	LoginController logCon;
+	
 	@PostMapping("/addtest")
-	public DiagnosticTest addTest(@RequestBody DiagnosticTest test) {
+	public DiagnosticTest addTest(@RequestBody DiagnosticTest test) throws ForBiddenException {
+		if(!logCon.loginStatus()) throw new ForBiddenException();
 		return testService.addTest(test);
 	}
 	@PutMapping("/updatetest")
-	public DiagnosticTest updateTest(@RequestBody DiagnosticTest test) throws DataNotFoundInDataBase {
+	public DiagnosticTest updateTest(@RequestBody DiagnosticTest test) throws DataNotFoundInDataBase, ForBiddenException {
+		if(!logCon.loginStatus()) throw new ForBiddenException();
 		return testService.updateTest(test);
 	}
 	@DeleteMapping("/removetest")
-	public DiagnosticTest removeTest(@RequestBody DiagnosticTest test) {
+	public DiagnosticTest removeTest(@RequestBody DiagnosticTest test) throws ForBiddenException {
+		if(!logCon.loginStatus()) throw new ForBiddenException();
 		return testService.removeTest(test);
 	}
 	@GetMapping("/viewalltest")
-	public List<DiagnosticTest> viewAllTest(){
+	public List<DiagnosticTest> viewAllTest() throws ForBiddenException{
+		if(!logCon.loginStatus()) throw new ForBiddenException();
 		return testService.viewAllTest();
 	}
 	@PostMapping("/addtestincenter")
-	public DiagnosticTest addTestInCenter(@RequestBody DiagnosticTest test,@RequestBody DiagnosticCenter center) {
+	public DiagnosticTest addTestInCenter(@RequestBody DiagnosticTest test,@RequestBody DiagnosticCenter center) throws ForBiddenException {
+		if(!logCon.loginStatus()) throw new ForBiddenException();
 		return testService.addTestInCenter( test, center);
 	}
 }
